requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.3
Django==3.2.5
django-cleanup==5.2.0
django-cors-headers==3.7.0
django-crispy-forms==1.10.0
django-excel-response==2.0.5
django-jsonview==2.0.0
et-xmlfile==1.1.0
idna==3.2
jdcal==1.4.1
openpyxl==2.4.11
Pillow==8.0.1
python-dotenv==0.15.0
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.3.0
XlsxWriter==3.0.1
